Update 3_contorno.f90
Ajustado da real 4 para real 8 na linha 1248

+++ SuLi-ICdz/3_contorno.f90
@@ -1245,7 +1245,7 @@ SUBROUTINE boundary_waves()
 	real(8),dimension(0:nx+1,0:ny1+1,0:nz+1) :: v1
 	real(8),dimension(0:nx+1,0:ny+1,0:nz1+1) :: w1
 
-	real(4),dimension(0:nx+1,0:ny+1,0:nz+1) :: ls1
+	real(8),dimension(0:nx+1,0:ny+1,0:nz+1) :: ls1
 
 
 	!Reference: Coubilla, 2015 (Thesis)
